Update Java and JVM versions in build.gradel.kts
The Java source compatibility and target compatibility were updated to version 17 in the license-aggregator-compose/build.gradle.kts. Moreover, the JVM target was also updated to "17". These updates were necessary in order to take advantage of the latest Java features and ensure that the code runs on the most recent JVM version.